Fix search table and marker remove, add dismiss default location btn, add separate weather icons layer
If applied, this commit will fix search table overload on repeated search button clicks.
It is ensured that only one location marker is present on the map in the same time.
Dismiss button is added in the location's marker popup. If set to default location, the location marker
automatically provides dismiss button popup (instead of Make default button popup).
Weather icons are put in the separate map layer with a checkbox view for turning the layer on and off in the upper
right corner layers menu tab in the user interface.